[api] organizations: more tests.
