+++ lib/http-users/organization/core.js
@@ -133,7 +133,10 @@ exports.resource = function (app) {
 
       org.members.push(username);
 
-      return Organization.update(org._id, { members: org.members }, callback);
+      return Organization.update(org._id, { members: org.members }, function (err) {
+        if (err) return callback(err);
+        return callback(null, org);
+      });
     });
   };
 
@@ -168,7 +171,10 @@ exports.resource = function (app) {
       owners: org.owners
     };
 
-    return Organization.update(org._id, data, callback);
+    return Organization.update(org._id, data, function (err) {
+      if (err) return callback(err);
+      return callback(null, org);
+    });
   };
 
   //
@@ -200,7 +206,10 @@ exports.resource = function (app) {
         owners: org.owners
       };
 
-      return Organization.update(org._id, data, callback);
+      return Organization.update(org._id, data, function (err) {
+        if (err) return callback(err);
+        return callback(null, org);
+      });
     });
   };
 
@@ -217,7 +226,20 @@ exports.resource = function (app) {
       return callback(new Error('Specified user is not an owner.'));
     }
 
-    return this.removeMember(username, callback);
+    // return org.removeMember(username, callback);
+
+    var i = org.owners.indexOf(username);
+    if (~i) {
+      if (org.owners.length === 1) {
+        return callback(new Error('Cannot remove the only owner.'));
+      }
+      org.owners.splice(i, 1);
+    }
+
+    return Organization.update(org._id, { owners: org.owners }, function (err) {
+      if (err) return callback(err);
+      return callback(null, org);
+    });
   };
 
   //
